requirements.txt
easydict==1.10
kaggle==1.5.15
opencv-python==4.7.0.72
pandas==1.5.3
Pillow==8.4.0
PyYAML==6.0
robustness==1.2.1
scipy==1.10.1
tensorboardX==2.6.1
torch==2.0.1
torchvision==0.15.2